added: authentication requirements for several endpoints

+++ part2/hbnb/app/api/v1/reviews.py
@@ -1,4 +1,5 @@
 from flask_restx import Namespace, Resource, fields
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from app.services import facade
 
 
@@ -21,20 +22,37 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def post(self):
+        current_user = get_jwt_identity()
         new_review = api.payload
-        if not new_review.get('text'):
-            return {'message': 'Text is required'}, 400
-        if int(new_review.get('rating')) < 1 or int(new_review.get('rating')) > 5:
-            return {'message': 'Rating must be between 1 and 5'}, 400
-        if not new_review.get('user_id'):
-            return {'message': 'User ID is required'}, 400
-        if not facade.get_user(new_review.get('user_id')):
-            return {'message': 'User does not exist'}, 400
-        if not new_review.get('place_id'):
-            return {'message': 'Place ID is required'}, 400
+        new_review['user_id'] = current_user['id']
+
         if not facade.get_place(new_review.get('place_id')):
             return {'message': 'Place does not exist'}, 400
+
+        place = facade.get_place(new_review.get('place_id'))
+        for review in place.reviews:
+            if review.user == current_user['id']:
+                return {'message': 'You have already reviewed this place'}, 400
+
+        if place.owner == new_review.get('user_id'):
+            return {'message': 'User cannot review their own place'}, 400
+
+        if not new_review.get('text'):
+            return {'message': 'Text is required'}, 400
+
+        if int(new_review.get('rating')) < 1 or int(new_review.get('rating')) > 5:
+            return {'message': 'Rating must be between 1 and 5'}, 400
+
+        if not new_review.get('user_id'):
+            return {'message': 'User ID is required'}, 400
+
+        if not facade.get_user(new_review.get('user_id')):
+            return {'message': 'User does not exist'}, 400
+
+        if not new_review.get('place_id'):
+            return {'message': 'Place ID is required'}, 400
 
         created_review = facade.create_review(new_review)
         return {
@@ -60,9 +78,16 @@ class ReviewResource(Resource):
     @api.response(200, 'Review details retrieved successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
+        current_user = get_jwt_identity()
         updated_data = api.payload
+
+        review = facade.get_review(review_id)
+        if review.user != current_user['id']:
+            return {"error": 'Unauthorized action'}, 403
+
         if not updated_data.get('text'):
             return {'message': 'Text is required'}, 400
         if int(updated_data.get('rating')) < 1 or int(updated_data.get('rating')) > 5:
@@ -85,15 +110,20 @@ class ReviewResource(Resource):
             "user_id": review.user,
             "place_id": review.place
         }, 200
-    
+
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @jwt_required()
     def delete(self, review_id):
         """Delete a review"""
         review = facade.get_review(review_id)
+        current_user = get_jwt_identity()
+
         if not review:
             return {'error': 'Review not found'}, 404
+
+        if review.user != current_user['id']:
+            return {"error": 'Unauthorized action'}, 403
+
         facade.delete_review(review_id)
         return {'message': 'Review deleted successfully'}, 200
-    
-
